feat: Webhooks optimize parallel processing
